+++ codec/src/main/java/io/netty/handler/codec/compression/ZlibDecoder.java
@@ -29,6 +29,7 @@ public abstract class ZlibDecoder extends ByteToMessageDecoder {
      * Maximum allowed size of the decompression buffer.
      */
     protected final int maxAllocation;
+    protected final boolean preferDirect = false;
 
     /**
      * Same as {@link #ZlibDecoder(int)} with maxAllocation = 0.
@@ -63,10 +64,10 @@ public abstract class ZlibDecoder extends ByteToMessageDecoder {
     protected ByteBuf prepareDecompressBuffer(ChannelHandlerContext ctx, ByteBuf buffer, int preferredSize) {
         if (buffer == null) {
             if (maxAllocation == 0) {
-                return ctx.alloc().heapBuffer(preferredSize);
+                return ctx.alloc().buffer(preferredSize);
             }
 
-            return ctx.alloc().heapBuffer(Math.min(preferredSize, maxAllocation), maxAllocation);
+            return ctx.alloc().buffer(Math.min(preferredSize, maxAllocation), maxAllocation);
         }
 
         // this always expands the buffer if possible, even if the expansion is less than preferredSize
